% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Psychological data}
\format{
## `example_data`
A data frame with 600 rows and 8 columns:
\describe{
  \item{locus_of_control}{Psychological Locus of Control}
  \item{self_concept}{Psychological Self Concept}
  \item{motivation}{Psychological Motivation}
  \item{read}{Academic Reading}
  \item{write}{Academic Writing}
  \item{math}{Academic Math}
  \item{science}{Academic Science}
  \item{female}{Binary flag for 1 being female.}
}
}
\source{
<https://stats.idre.ucla.edu/stat/data/mmreg.csv>
}
\usage{
example_data
}
\description{
A dataset from UCLA Statistical Methods and Data Analytics about
investigating the associations between psychological measures and academic achievement measures.
}
\keyword{datasets}
