% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqcanon-package.R
\docType{package}
\name{rqcanon-package}
\alias{rqcanon}
\alias{rqcanon-package}
\title{rqcanon: Canonical Quantile Regression}
\description{
A quantile regression method for multivariate data to find linear combinations of explanatory and response variables generalizing canonical correlations. The package consists of functions, rqcan() for fitting the coefficients, and summary.rqcan(), which calls a bootstrap R-function. For details, see the help files for rqcan() and summary.rqcan(), and the reference: Portnoy (2022) <doi:10.1016/j.jmva.2022.105071>.
}
\author{
\strong{Maintainer}: Stephen Portnoy \email{sportnoy@illinois.edu}

Other contributors:
\itemize{
  \item Isa Mostachetti [compiler]
  \item Daniel Taylor-Rodriguez [reviewer]
}

}
\keyword{internal}
